refactor: 💡 rebuild directory
