+++ src/components/functional/Form/index.ts
@@ -1,3 +1,4 @@
 export * from './Form'
 export * from './FieldWrapper'
 export * from './InputField'
+export * from './schema'
